// src/app/shared/models/food.ts
export interface Food {
  img: string;
  title: string;
  name: string;
  time: string;
  rating: string;
  ratingClass: string;
  additives: string;
  pg: string;
  fat: string;
  sugar: string;
  calories: string;
  salt: string;
  Description: string;
  storage: string;
  recommendation: string;
  reminder: string;
}